Modularise getting repos, provide explicit comments and save profile.
