+++ server/profiling/index.js
@@ -1,89 +1,59 @@
-// Libraries
-const axios = require('axios');
-
-// lodash functions
-const map = require('lodash/map');
-const forEach = require('lodash/forEach');
-const filter = require('lodash/filter');
+/*
+ *  PROFILING
+ * 
+ *  This module attempts to build profile users by experience
+ *  by scraping info from GitHub's API.
+ * 
+ *  This profile is saved as JSON on the user's node in the db and
+ *  used to compare users by finding the absolute difference
+ *  between their relative experience.
+ * 
+ *  There's a lot of room to add features here, as GitHub's API
+ *  is very generous with information. It is less generous, however,
+ *  with allowances: we added some very cool features but quickly found
+ *  that doing all the necessary API queries was blasting our rate limit.
+ * 
+ *  NOTES:
+ *  --lodash
+ *  lodash functions are used here, as there's an awful lot of repeated
+ *  iteration over arrays, and for that lodash is generally the fastest.
+ * 
+ *  --Style
+ *  This is generally written (with some notable exceptions--e.g. the mutation of profile)
+ *  in quite a functional style. This is to make it easier to add new functionality into
+ *  the chain.
+ * 
+ *  Each function takes some information does some processing and returns new information
+ *  for the next function. We'd advise you stick with this layout, but you may wish to change
+ *  the object destructuring simply to positional parameters. We used it as a convenience,
+ *  allowing us to add features and information (and also drop them) quickly, without
+ *  worrying about how other function's might view those changes.
+ */
 
 // Other server modules
 const db = require('../db');
-// Other server modules for profiling
+// Other profiling modules
 const getUserLanguages = require('./languages');
 const getUserRepoStats = require('./repoStats');
+const getUserRepos = require('./repos');
 
-// Gets all repos that a user either created or has committed to
-exports.getUserRepos = function getUserRepos(ghId) {
-  // First get user info from the db
+// Takes a profile and saves it to the db
+const saveUserProfile = function saveUserProfile({ username, OAuthToken, repos, profile }) {
   const dbSession = db.driver.session();
-  return dbSession.run(`
-    MATCH (user:User {ghId: ${ ghId }}) RETURN user
+  dbSession.run(`
+    MATCH (user:User {OAuthToken: '${OAuthToken}'})
+    SET user.profile = '${JSON.stringify(profile).replace('\'', '\\\'')}'
   `)
-    .then((res) => {
-      // Then use that info to get the users repos from GitHub
-      const { OAuthToken, username } = new db.models.User(res.records[0].get('user'));
-      return axios.get('https://api.github.com/user/repos', {
-        params: {
-          access_token: OAuthToken,
-          per_page: 100,
-        }
-      })
-        .then(res => ({ OAuthToken, username, res }));
-    })
-    .then(({ OAuthToken, username, res }) =>
-      // Then extract the useful information from those repos
-      // Here we're interested in the owner, name, whether it's forked 
-      // and the url for its language stats
-      ({
-        OAuthToken,
-        username,
-        repos: map(res.data, ({ name, owner, languages_url, fork, forks_count, stargazers_count, watchers_count }) => 
-          ({ name, owner: owner.login, languages_url, fork, forks_count, stargazers_count, watchers_count }))
-      })
-    )
-    .then(({ username, OAuthToken, repos }) => {
-      // Then check if they're forks. 
-      // If so, find out whether or not the user has actually committed to that repo.
-      const commitPromises = [];
-      forEach(repos, (repo) => {
-        if (repo.fork) {
-          // It's a fork. so query for commits
-          commitPromises.push(axios.get(`https://api.github.com/repos/${ username }/${ repo.name }/commits`, {
-            params: {
-              access_token: OAuthToken,
-              author: username,
-            },
-          })
-            .then((res) => {
-              // Check for commits
-              repo.include = res.data.length > 0;
-            })
-          );
-        } else {
-          // Not a fork.
-          repo.include = true;
-        }
-      });
-      return Promise.all(commitPromises)
-        .then(() => ({ username, OAuthToken, repos }));
-    })
-    .then(({ OAuthToken, username, repos }) => 
-      // Finally filter by the include property
-      ({
-        OAuthToken,
-        username,
-        repos: filter(repos, repo => repo.include),
-        profile: {}
-      })
-    );
+  .then(() => { username, OAuthToken, repos, profile })
+  .catch(console.error);
 };
 
-// This is a testing function which uses the above chain
-// and one from the languages module to show a user's most-used languages
-exports.displayUserLanguages = function gl(ghId) {
-  exports.getUserRepos(ghId)
-    // .then(getUserLanguages)
+// This function chains together all the separate profiling functionality
+// to create a user profile and save it to their node on the db.
+exports.buildUserProfile = function buildUserProfile(ghId) {
+  getUserRepos(ghId)
+    .then(getUserLanguages)
     .then(getUserRepoStats)
-    .then(console.log)
+    .then(saveUserProfile)
     .catch(console.error);
 }
